+++ CHANGELOG.md
@@ -47,6 +47,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+#### Core
+- ``..`` pattern cannot match empty directory.
+
 #### Completer
 - always set new ``COMPREPLY`` variable even if completion result is empty